refactor: Modify SignIn View
